pyproject.toml
[project]
name = "auto-hedge"
version = "0.1.0"
description = "Automatic open contr position for futures trading"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "binance-futures-connector>=4.1.0",
    "pydantic-settings>=2.7.1",
]
